Guard episode and stream tables without foreign keys

+++ app/database/migrations/2024_08_18_000000_create_watch_progress_table.php
@@ -10,7 +10,7 @@ return new class extends Migration {
         Schema::create('watch_progress', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('anime_id');
+            $table->unsignedSmallInteger('anime_id');
             $table->unsignedInteger('episode_number');
             $table->timestamps();
